zephyr_helpers.py: add a CLI
Provide a command-line interface for analyzing a Zephyr repository.
Nothing about this behavior is Foundries.io specific, but the behavior
of sauce tags turns out to be kind of confusing and should be fixed up
in a future patch.

We do change the behavior of the library itself to return better
errors when the upstream or downstream refs are invalid, and we switch
from printing to logging as needed.

For now, support listing:

- printing out of tree patches
- printing likely merged patches

For ease of scripting, support printing in JSON or "prettified" JSON.

Additionally allow the user to provide options via a configuration
file. This avoids tedious repetition for required arguments and allows
user-provided "defaults" for other options.

By default, we look in ~/.config/zephyr_helpers.config for a
configuration file. Additional files can be specified on the command
line.

Add a "template" configuration file as well for people to start out
with.

Signed-off-by: Marti Bolivar <marti@foundries.io>